+++ src/main/admin-page/src/AxiosRequest/Axios.ts
@@ -2,7 +2,7 @@ import axios from "axios";
 
 const API_URL = "http://localhost:8000";
 
-const request = axios.create({
+export const request = axios.create({
     baseURL: API_URL
 });
 
@@ -40,31 +40,3 @@ export const LoginAxios = (loginData: { id: string, password: string }) => {
     });
 };
 
-export const getAnnouncement = async (params: undefined | null | { sortingColumn: string, sortingMethod: 'ASC' | 'DESC' }, page: number = 1) => {
-    const response = await request.get('/announcement', {
-        params: {
-            pageNum: page,
-            orderCriteria: params?.sortingColumn,
-            orderMethod: params?.sortingMethod,
-        },
-        headers: { 'Content-Type': 'application/json' }
-    });
-
-    return response;
-};
-
-export const getAnnouncementById = async (announcementId: number) => {
-    const response = await request.get(`/announcement/${announcementId}`, {
-        headers: { 'Content-Type': 'application/json' }
-    });
-
-    return response;
-};
-
-export const getBlockUser = async () => {
-    const response = await request.get('/admin/blockuser', {
-        headers: { 'Content-Type': 'application/json' }
-    });
-
-    return response;
-};
