+++ src/main/admin-page/src/AxiosRequest/BlockInfoApi.ts
@@ -1,0 +1,9 @@
+import { request } from "./Axios";
+
+export const getBlockUser = async () => {
+    const response = await request.get('/admin/blockuser', {
+        headers: { 'Content-Type': 'application/json' }
+    });
+
+    return response;
+};
